+++ CHANGELOG.md
@@ -2,7 +2,8 @@ v0.2.0 (in development)
 -----------------------
 - Require the `port` field of `SMTPSender` to be non-negative
 - Mark `Sender` as `runtime_checkable` and export it
-- Gave the `outgoing` command `--section` and `--no-section` options
+- Gave the `outgoing` command `--section`, `--no-section`, and `--log-level`
+  options
 - Added logging to built-in sender classes
 
 v0.1.0 (2021-03-06)